Getting token based authentication to work using allauth (had to rename account app into user_account)
