--- account/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-
-class AccountConfig(AppConfig):
-    name = 'account'
-
-#     def ready(self):
-#         from . import signals
