+++ requirements.txt
@@ -7,3 +7,4 @@ django-rest-auth==0.7.0
 django-guardian==1.4.6
 django-treebeard==4.1.0
 django-filter==1.0.1
+Pillow==4.0.0
