+++ user_account/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class USerAccountConfig(AppConfig):
+    name = 'user_account'
+
+#     def ready(self):
+#         from . import signals
