+++ user_account/authentication/urls.py
@@ -1,0 +1,33 @@
+from django.conf.urls import url
+
+from user_account.authentication.views import UserDetailsView, PasswordResetView,\
+    PasswordResetConfirmView, LoginView, LogoutView, PasswordChangeView
+
+
+urlpatterns = [
+    # URLs that do not require a session or valid token
+    url(r'^password/reset/$',
+        PasswordResetView.as_view(),
+        name='rest_password_reset'),
+
+    url(r'^password/reset/confirm/$',
+        PasswordResetConfirmView.as_view(),
+        name='rest_password_reset_confirm'),
+
+    url(r'^login/$',
+        LoginView.as_view(),
+        name='rest_login'),
+
+    # URLs that require a user to be logged in with a valid session / token.
+    url(r'^logout/$',
+        LogoutView.as_view(),
+        name='rest_logout'),
+
+    url(r'^user/$',
+        UserDetailsView.as_view(),
+        name='rest_user_details'),
+
+    url(r'^password/change/$',
+        PasswordChangeView.as_view(),
+        name='rest_password_change'),
+]
